fix user role update
